+++ common/constants/custom_panels.ts
@@ -8,7 +8,6 @@ import { v4 as uuidv4 } from 'uuid';
 export const CUSTOM_PANELS_API_PREFIX = '/api/observability/operational_panels';
 export const CUSTOM_PANELS_DOCUMENTATION_URL =
   'https://opensearch.org/docs/latest/observability-plugin/operational-panels/';
-export const OPENSEARCH_DOCUMENTATION_URL = 'https://opensearch.org/docs/latest';
 export const CREATE_PANEL_MESSAGE =
   'Enter a name to describe the purpose of this Observability Dashboard.';
