+++ common/constants/explorer.ts
@@ -38,16 +38,6 @@ export const CREATE_TAB_PARAM = {
 } as const;
 export const HAS_SAVED_TIMESTAMP = 'hasSavedTimestamp';
 export const FILTER_OPTIONS = ['Visualization', 'Query', 'Metric'];
-export const ASSET_FILTER_OPTIONS = ['index-pattern', 'search', 'visualization', 'dashboard'];
-export const INTEGRATION_TEMPLATE_OPTIONS = ['nginx', 'aws_elb'];
-export const INTEGRATION_CATEOGRY_OPTIONS = [
-  'communication',
-  'http',
-  'cloud',
-  'aws_elb',
-  'container',
-  'logs',
-];
 export const SAVED_QUERY = 'savedQuery';
 export const SAVED_VISUALIZATION = 'savedVisualization';
 export const SAVED_OBJECT_ID = 'savedObjectId';
