+++ common/constants/integrations.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export const OPENSEARCH_DOCUMENTATION_URL = 'https://opensearch.org/docs/latest';
+export const ASSET_FILTER_OPTIONS = ['index-pattern', 'search', 'visualization', 'dashboard'];
+export const INTEGRATION_TEMPLATE_OPTIONS = ['nginx', 'aws_elb'];
+export const INTEGRATION_CATEOGRY_OPTIONS = [
+  'communication',
+  'http',
+  'cloud',
+  'aws_elb',
+  'container',
+  'logs',
+];
