+++ public/components/integrations/components/__tests__/mapping_validation.test.ts
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import {
   doTypeValidation,
   doNestedPropertyValidation,
