+++ public/components/integrations/components/added_integration.tsx
@@ -29,7 +29,7 @@ import {
 import React, { useEffect, useState } from 'react';
 import _ from 'lodash';
 import { PanelTitle } from '../../trace_analytics/components/common/helper_functions';
-import { ASSET_FILTER_OPTIONS, FILTER_OPTIONS } from '../../../../common/constants/explorer';
+import { ASSET_FILTER_OPTIONS } from '../../../../common/constants/integrations';
 import { INTEGRATIONS_BASE, OBSERVABILITY_BASE } from '../../../../common/constants/shared';
 import { DeleteModal } from '../../common/helpers/delete_modal';
 import { AddedIntegrationProps } from './integration_types';
@@ -271,83 +271,6 @@ export function AddedIntegration(props: AddedIntegrationProps) {
           itemId="id"
           loading={false}
           items={assets}
-          columns={tableColumns}
-          pagination={{
-            initialPageSize: 10,
-            pageSizeOptions: [5, 10, 15],
-          }}
-          search={search}
-        />
-      </EuiPanel>
-    );
-  }
-
-  function AddedIntegrationFields(fieldProps: any) {
-    const data = fieldProps.data?.fields || [];
-
-    const search = {
-      box: {
-        incremental: true,
-      },
-      filters: [
-        {
-          type: 'field_value_selection',
-          field: 'type',
-          name: 'Type',
-          multiSelect: false,
-          options: FILTER_OPTIONS.map((i) => ({
-            value: i,
-            name: i,
-            view: i,
-          })),
-        },
-      ],
-    };
-
-    const tableColumns = [
-      {
-        field: 'name',
-        name: 'Name',
-        sortable: true,
-        truncateText: true,
-        render: (value, record) => (
-          <EuiText data-test-subj={`${record.name}IntegrationLink`}>
-            {_.truncate(record.description, { length: 100 })}
-          </EuiText>
-        ),
-      },
-      {
-        field: 'type',
-        name: 'Type',
-        sortable: true,
-        truncateText: true,
-        render: (value, record) => (
-          <EuiText data-test-subj={`${record.type}IntegrationDescription`}>
-            {_.truncate(record.type, { length: 100 })}
-          </EuiText>
-        ),
-      },
-      {
-        field: 'category',
-        name: 'Category',
-        sortable: true,
-        truncateText: true,
-        render: (value, record) => (
-          <EuiText data-test-subj={`${record.type}IntegrationDescription`}>
-            {_.truncate(record.category, { length: 100 })}
-          </EuiText>
-        ),
-      },
-    ] as Array<EuiTableFieldDataColumnType<any>>;
-
-    return (
-      <EuiPanel>
-        <PanelTitle title={'Integration Fields'} />
-        <EuiSpacer size="l" />
-        <EuiInMemoryTable
-          itemId="id"
-          loading={false}
-          items={data}
           columns={tableColumns}
           pagination={{
             initialPageSize: 10,
