+++ public/components/integrations/components/added_integration_table.tsx
@@ -21,7 +21,7 @@ import { AddedIntegrationsTableProps } from './added_integration_overview_page';
 import {
   ASSET_FILTER_OPTIONS,
   INTEGRATION_TEMPLATE_OPTIONS,
-} from '../../../../common/constants/explorer';
+} from '../../../../common/constants/integrations';
 import { DeleteModal } from '../../../../public/components/common/helpers/delete_modal';
 import { INTEGRATIONS_BASE } from '../../../../common/constants/shared';
 import { useToast } from '../../../../public/components/common/toast';
@@ -154,17 +154,8 @@ export function AddedIntegrationsTable(props: AddedIntegrationsTableProps) {
             pageSizeOptions: [5, 10, 15],
           }}
           search={search}
-          // sorting={{
-          //   sort: {
-          //     field: 'dateModified',
-          //     direction: 'desc',
-          //   },
-          // }}
           allowNeutralSort={false}
           isSelectable={true}
-          // selection={{
-          //   onSelectionChange: (items) => setSelectedApplications(items),
-          // }}
         />
       ) : (
         <>
