+++ public/components/integrations/components/available_integration_card_view.tsx
@@ -99,14 +99,9 @@ export function AvailableIntegrationsCardView(props: AvailableIntegrationsCardVi
             isClearable={false}
             placeholder="Search..."
             data-test-subj="search-bar-input-box"
-            // value={query}
-            // onClick = {() => query.current.}
             onChange={(e) => {
               props.setQuery(e.target.value);
-              // query.current
-              // setGlobalQuery(e.target.value);
             }}
-            // onSearch={props.refresh}
           />
         </EuiFlexItem>
         <EuiFlexItem grow={false}>{props.renderCateogryFilters()}</EuiFlexItem>
