+++ public/components/integrations/components/integration_header.tsx
@@ -15,7 +15,7 @@ import {
 } from '@elastic/eui';
 import _ from 'lodash';
 import React, { useEffect, useState } from 'react';
-import { OPENSEARCH_DOCUMENTATION_URL } from '../../../../common/constants/custom_panels';
+import { OPENSEARCH_DOCUMENTATION_URL } from '../../../../common/constants/integrations';
 
 export function IntegrationHeader() {
   const tabs = [
