+++ public/components/integrations/home.tsx
@@ -11,7 +11,6 @@ import { Integration } from './components/integration';
 import { TraceAnalyticsCoreDeps } from '../trace_analytics/home';
 import { ChromeBreadcrumb } from '../../../../../src/core/public';
 import { AvailableIntegrationOverviewPage } from './components/available_integration_overview_page';
-import { Sidebar } from './components/integration_side_nav';
 import { AddedIntegrationOverviewPage } from './components/added_integration_overview_page';
 import { AddedIntegration } from './components/added_integration';
