+++ server/adaptors/integrations/__data__/repository/aws_elb/aws_elb-1.0.0.json
@@ -6,7 +6,7 @@
     "license": "Apache-2.0",
     "type": "logs",
     "author": "OpenSearch",
-    "sourceUrl": "https://github.com/YANG-DB/observability/tree/aws_alb_integration",
+    "sourceUrl": "https://github.com/opensearch-project/opensearch-catalog/blob/main/schema/observability/logs/aws_elb.mapping",
     "statics": {
         "logo": {
             "annotation": "ELB Logo",
