+++ server/adaptors/integrations/__data__/repository/nginx/nginx-1.0.0.json
@@ -7,7 +7,7 @@
     "type": "logs",
     "link": "https://www.nginx.com/",
     "author": "OpenSearch",
-    "sourceUrl": "https://github.com/opensearch-project/observability/pull/1493/files",
+    "sourceUrl": "https://github.com/opensearch-project/observability/tree/2.x/integrations/nginx",
     "statics": {
         "logo": {
             "annotation": "NginX Logo",
